Moved WanIp class definition to a seperate file
