+++ archey/archey.py
@@ -7,8 +7,7 @@ import sys
 
 from enum import Enum
 from glob import glob
-from subprocess import CalledProcessError, DEVNULL, \
- TimeoutExpired, check_output
+from subprocess import CalledProcessError, DEVNULL, check_output
 
 import distro
 
@@ -22,6 +21,7 @@ from disk import Disk
 from ram import RAM
 from cpu import CPU
 from lan_ip import LanIp
+from wan_ip import WanIp
 from constants import (
     COLOR_DICT,
     DE_DICT,
@@ -286,80 +286,6 @@ class GPU:
         self.value = gpuinfo
 
 
-class WanIp:
-    def __init__(self):
-        # IPv6 address retrieval (unless the user doesn't want it).
-        if CONFIG.get('ip_settings')['wan_ip_v6_support']:
-            try:
-                ipv6_value = check_output(
-                    [
-                        'dig', '+short', '-6', 'AAAA', 'myip.opendns.com',
-                        '@resolver1.ipv6-sandbox.opendns.com'
-                    ],
-                    timeout=CONFIG.get('timeout')['ipv6_detection'],
-                    stderr=DEVNULL, universal_newlines=True
-                ).rstrip()
-
-            except (FileNotFoundError, TimeoutExpired, CalledProcessError):
-                try:
-                    ipv6_value = check_output(
-                        [
-                            'wget', '-q6O-', 'https://v6.ident.me/'
-                        ],
-                        timeout=CONFIG.get('timeout')['ipv6_detection'],
-                        universal_newlines=True
-                    )
-
-                except (CalledProcessError, TimeoutExpired):
-                    # It looks like this user doesn't have any IPv6 address...
-                    # ... or is not connected to Internet.
-                    ipv6_value = None
-
-                except FileNotFoundError:
-                    ipv6_value = None
-                    print('Warning: `wget` has not been found on your system.',
-                          file=sys.stderr)
-
-        else:
-            ipv6_value = None
-
-        # IPv4 addresses retrieval (anyway).
-        try:
-            ipv4_value = check_output(
-                [
-                    'dig', '+short', '-4', 'A', 'myip.opendns.com',
-                    '@resolver1.opendns.com'
-                ],
-                timeout=CONFIG.get('timeout')['ipv4_detection'],
-                stderr=DEVNULL, universal_newlines=True
-            ).rstrip()
-
-        except (FileNotFoundError, TimeoutExpired, CalledProcessError):
-            try:
-                ipv4_value = check_output(
-                    [
-                        'wget', '-q4O-', 'https://v4.ident.me/'
-                    ],
-                    timeout=CONFIG.get('timeout')['ipv4_detection'],
-                    universal_newlines=True
-                )
-
-            except (CalledProcessError, TimeoutExpired):
-                # This user looks not connected to Internet...
-                ipv4_value = None
-
-            except FileNotFoundError:
-                ipv4_value = None
-                # If statement so as to not print the same message twice.
-                if not CONFIG.get('ip_settings')['wan_ip_v6_support']:
-                    print('Warning: `wget` has not been found on your system.',
-                          file=sys.stderr)
-
-        self.value = ', '.join(
-            filter(None, (ipv4_value, ipv6_value))
-        ) or CONFIG.get('default_strings')['no_address']
-
-
 # ----------- Classes Index ----------- #
 
 class Classes(Enum):
@@ -386,7 +312,15 @@ class Classes(Enum):
             'no_address': CONFIG.get('default_strings')['no_address']
         }
     }
-    WAN_IP = {'class': WanIp}
+    WAN_IP = {
+        'class': WanIp,
+        'kwargs': {
+            'ipv6_support': CONFIG.get('ip_settings')['wan_ip_v6_support'],
+            'ipv6_timeout': CONFIG.get('timeout')['ipv6_detection'],
+            'ipv4_timeout': CONFIG.get('timeout')['ipv4_detection'],
+            'no_address': CONFIG.get('default_strings')['no_address']
+        }
+    }
 
 
 # ---------------- Main --------------- #
