+++ archey/wan_ip.py
@@ -1,0 +1,85 @@
+import sys
+from subprocess import (
+    check_output,
+    DEVNULL,
+    TimeoutExpired,
+    CalledProcessError
+)
+
+
+class WanIp:
+    def __init__(self,
+                 ipv6_support=False,
+                 ipv6_timeout=0,
+                 ipv4_timeout=0,
+                 no_address=None):
+        # IPv6 address retrieval (unless the user doesn't want it).
+        if ipv6_support:
+            try:
+                ipv6_value = check_output(
+                    [
+                        'dig', '+short', '-6', 'AAAA', 'myip.opendns.com',
+                        '@resolver1.ipv6-sandbox.opendns.com'
+                    ],
+                    ipv6_timeout,
+                    stderr=DEVNULL, universal_newlines=True
+                ).rstrip()
+
+            except (FileNotFoundError, TimeoutExpired, CalledProcessError):
+                try:
+                    ipv6_value = check_output(
+                        [
+                            'wget', '-q6O-', 'https://v6.ident.me/'
+                        ],
+                        ipv6_timeout,
+                        universal_newlines=True
+                    )
+
+                except (CalledProcessError, TimeoutExpired):
+                    # It looks like this user doesn't have any IPv6 address...
+                    # ... or is not connected to Internet.
+                    ipv6_value = None
+
+                except FileNotFoundError:
+                    ipv6_value = None
+                    print('Warning: `wget` has not been found on your system.',
+                          file=sys.stderr)
+
+        else:
+            ipv6_value = None
+
+        # IPv4 addresses retrieval (anyway).
+        try:
+            ipv4_value = check_output(
+                [
+                    'dig', '+short', '-4', 'A', 'myip.opendns.com',
+                    '@resolver1.opendns.com'
+                ],
+                ipv4_timeout,
+                stderr=DEVNULL, universal_newlines=True
+            ).rstrip()
+
+        except (FileNotFoundError, TimeoutExpired, CalledProcessError):
+            try:
+                ipv4_value = check_output(
+                    [
+                        'wget', '-q4O-', 'https://v4.ident.me/'
+                    ],
+                    ipv4_timeout,
+                    universal_newlines=True
+                )
+
+            except (CalledProcessError, TimeoutExpired):
+                # This user looks not connected to Internet...
+                ipv4_value = None
+
+            except FileNotFoundError:
+                ipv4_value = None
+                # If statement so as to not print the same message twice.
+                if not ipv6_support:
+                    print('Warning: `wget` has not been found on your system.',
+                          file=sys.stderr)
+
+        self.value = ', '.join(
+            filter(None, (ipv4_value, ipv6_value))
+        ) or no_address
